add shared.c, win32_get_home
change sig. of ecalloc

+++ functions.h
@@ -4,8 +4,10 @@
     This replaces the old perl-generated funcs.h
 */
 
+#include <stddef.h>
+
 char * save (const char *s);
-void* ecalloc (int count, unsigned int size);
+void* ecalloc (size_t count, size_t size);
 char * skipsp (char *s);
 int sprefix (char *ps, char *s, int uppercase);
 void quit (int status);
